@@ -1,17 +1,20 @@
 buildscript {
     repositories {
-        jcenter()
         google()
+        mavenCentral()
+        maven { url "https://jitpack.io" }
+        maven {
+            url "https://plugins.gradle.org/m2/"
+        }
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.1'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
+        classpath 'com.android.tools.build:gradle:8.5.2'
     }
 }
 
 apply plugin: 'com.android.library'
+apply plugin: 'maven-publish'
 
 ext {
     bintrayRepo = 'maven'
@@ -38,13 +41,14 @@ ext {
 }
 
 android {
-    compileSdkVersion 26
+    compileSdk 34
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 26
+        minSdkVersion 21
+        targetSdkVersion 34
         versionCode 1
         versionName "1.9.0"
+        buildConfigField "String", "VERSION_NAME", "\"${versionName}\""
     }
     buildTypes {
         release {
@@ -53,23 +57,31 @@ android {
         }
     }
 
-    sourceSets{
+    sourceSets {
         main {
             jni.srcDirs = []
             jniLibs.srcDir 'src/main/libs'
         }
     }
+    buildFeatures {
+        buildConfig = true
+    }
+    namespace = "com.shockwave.pdfium"
 }
 
 repositories {
     google()
-    jcenter()
+    mavenCentral()
+    maven { url "https://jitpack.io" }
+    maven {
+        url "https://plugins.gradle.org/m2/"
+    }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:support-v4:26.1.0'
+    implementation 'androidx.core:core:1.13.1'
 }
 
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
+//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
+//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
